[WIP] Implement JWK set parser converting JWKs into RSA keys.
JWT implementation can handle multiple type of keys for signing and
validating JSON web tokens. IETF defined a JSON representation of those
keys in
https://tools.ietf.org/html/rfc7517. This review implements this RFC.

This commit adds a parser to convert a JSON into a JWK set containing
RSA keys compatible with the implementation of RS256 in the JWT library.

The parser only supports parsing of  RSA keys for now but it can support
multiple types of keys such as elliptic curve keys and symmetric keys
as documented in the JWA RFC.